requirements.txt
streamlit==1.31.0
google-generativeai==0.3.1
Pillow==10.0.0
python-dotenv==1.0.0
pandas==2.1.0